app: refactor, post -> feed-tile
