+++ packages/app/src/pages/child-feed/index.tsx
@@ -19,7 +19,7 @@ import { ChildStackParamList } from "../../navigation/child-stack-navigator";
 import Icons, { Icon } from "../../components/icon";
 import * as Types from "../../types";
 import colors from "../../colors";
-import Post from "../../components/post";
+import FeedTile from "../../components/feed-tile";
 
 const fullWidth = Dimensions.get("window").width;
 
@@ -42,7 +42,7 @@ const Feed = () => {
   const feed = category.feed;
 
   const renderItem = ({ item }: { item: Types.IPostEdge }) => {
-    return <Post post={item} />;
+    return <FeedTile post={item} />;
   };
 
   const curveRendered = (
